Add sim speed/quality benchmark with sub-second time limit support
Change BAI time_limit_seconds from uint64_t to double to support
sub-second time limits (e.g. 0.1s per turn). Add on-demand "benchsim"
test that plays game pairs of 2-ply vs 3-ply sim at 0.1s and 1.0s per
turn, tracking spread, per-player node counts, and time overruns.
Uses static eval when bag is empty.

+++ src/def/bai_defs.h
@@ -26,7 +26,7 @@ typedef struct BAIOptions {
   double delta;
   uint64_t sample_limit;
   uint64_t sample_minimum;
-  uint64_t time_limit_seconds;
+  double time_limit_seconds;
   int num_threads;
   double cutoff;
 } BAIOptions;
